FollowRepo is now functional and system works
