+++ src/Chirp.Infrastructure/Repositories/CheepRepository.cs
@@ -57,26 +57,6 @@ public class CheepRepository : ICheepRepository
         var result = await pageQuery.ToListAsync();
         return result;
     }
-    public async Task<List<AuthorDTO>> GetFollowingDTO(string user)
-    {
-        // Get the list of authors the user is following
-        var followingList = await GetFollowingList(user);
-
-        // Map each Author to an AuthorDTO
-        return followingList.Select(author => new AuthorDTO
-        {
-            UserName = author.UserName,
-            DisplayName = string.IsNullOrWhiteSpace(author.DisplayName) ? author.UserName : author.DisplayName,
-            FollowingList = new List<Author>() // Optional: Populate this if needed
-        }).ToList();
-    }
-
-
-    public async Task<List<Author>> GetFollowingList(string user)
-    {
-        return await _dbContext.Authors.Where(a => a.UserName == user).Select(a => a.FollowingList)
-            .FirstOrDefaultAsync() ?? new List<Author>();
-    }
 
     public async Task<AuthorDTO?> GetAuthorByName(string name)
     {
@@ -86,18 +66,6 @@ public class CheepRepository : ICheepRepository
             return null;
         }
         return AuthorDTO.fromAuthor(author);
-    }
-
-    public async Task<Author> GetAuthor(string name)
-    {
-        var author =  _dbContext.Authors.SingleOrDefault(a => a.UserName == name);
-
-        if (author == null)
-        {
-            throw new KeyNotFoundException($"No author with name {name} was found.");
-        }
-
-        return author;
     }
 
     public Author GetAuthorByEmail(string name)
@@ -166,74 +134,6 @@ public class CheepRepository : ICheepRepository
             });
 
         return await bioQuery.FirstOrDefaultAsync();
-    }
-
-    public async Task AddFollow(string user, string userFollowed)
-    {
-            var author = await GetAuthor(user);
-            var authorFollowed = await GetAuthor(userFollowed);
-            if (author.FollowingList.All(a => a.UserName != userFollowed))
-            {
-                author.FollowingList.Add(authorFollowed);
-                await _dbContext.SaveChangesAsync();
-
-                var followedUsers = string.Join(", ", author.FollowingList.Select(a => a.UserName));
-            }
-    }
-
-    public async Task<bool> IsFollowing(string user, string userFollowed)
-    {
-        var list = await _dbContext.Authors.Where(a => a.UserName == user).Select(a => a.FollowingList)
-            .FirstOrDefaultAsync();
-
-        if (list != null && list.Any(a => a.UserName == userFollowed))
-        {
-            return true;
-        }
-
-        return false;
-    }
-
-    public async Task Unfollow(string user, string userUnfollowed)
-    {
-        var authorUnfollowed = await GetAuthor(userUnfollowed);
-            _dbContext.Authors.Where(a =>a.UserName == user)
-                .Include(a => a.FollowingList).FirstOrDefault()!.FollowingList.Remove(authorUnfollowed);
-
-            await _dbContext.SaveChangesAsync();
-        }
-
-    public async Task<List<CheepDTO>> GetCheepsFromFollowing(int pageNumber, string username)
-    {
-        var lowerBound = (pageNumber - 1) * _pageSize;
-
-        var followingUsernames = (await GetFollowingList(username))
-            ?.Select(a => a.UserName)
-            .ToList();
-        //Skal ikke bruges længere, da der stadig skal returneres sine egne cheeps, selvom "jeg" ikke følger nogle
-        /*
-        if (followingUsernames == null || !followingUsernames.Any())
-        {
-            return new List<CheepDTO>();
-        }
-        */
-        var pageQuery = (from cheep in _dbContext.Cheeps
-                where followingUsernames.Contains(cheep.Author.UserName) || cheep.Author.UserName == username
-                orderby cheep.TimeStamp descending
-                select cheep)
-            .Include(c => c.Author)
-            .Skip(lowerBound)
-            .Take(_pageSize)
-            .Select(cheep => new CheepDTO
-            {
-                Author = AuthorDTO.fromAuthor(cheep.Author),
-                Text = cheep.Text,
-                TimeStamp = cheep.TimeStamp.ToString("MM/dd/yy HH:mm:ss", System.Globalization.CultureInfo.InvariantCulture)
-            });
-
-        // Hent og returner resultatet
-        var result = await pageQuery.ToListAsync();
-        return result;
     }
 
     public async Task DeleteAuthor(AuthorDTO authorDTO)
