+++ src/Chirp.Infrastructure/Services/FollowService.cs
@@ -15,49 +15,49 @@ public interface IFollowService
 public class FollowService : IFollowService
 {
 
-    private readonly CheepRepository _cheepRepository;
+    private readonly FollowRepository _followRepository;
 
-    public FollowService(CheepRepository followRepository)
+    public FollowService(FollowRepository followRepository)
     {
-        _cheepRepository = followRepository;
+        _followRepository = followRepository;
     }
 
     public async Task FollowAuthor(string followerId, string followeeId)
     {
         // Check if already following to avoid duplication
-        if (!await _cheepRepository.IsFollowing(followerId, followeeId))
+        if (!await _followRepository.IsFollowing(followerId, followeeId))
         {
-            await _cheepRepository.AddFollow(followerId, followeeId);
+            await _followRepository.AddFollow(followerId, followeeId);
         }
     }
 
     public async Task UnfollowAuthor(string followerId, string followeeId)
     {
-        if (await _cheepRepository.IsFollowing(followerId, followeeId))
+        if (await _followRepository.IsFollowing(followerId, followeeId))
         {
-            await _cheepRepository.Unfollow(followerId, followeeId);
+            await _followRepository.Unfollow(followerId, followeeId);
         }
     }
 
     public async Task<bool> IsFollowing(string followerId, string followeeId)
     {
-        return await _cheepRepository.IsFollowing(followerId, followeeId);
+        return await _followRepository.IsFollowing(followerId, followeeId);
     }
 
     public List<AuthorDTO> GetFollowing(string username)
     {
-        var authorlist = _cheepRepository.GetFollowingDTO(username).Result;
+        var authorlist = _followRepository.GetFollowingDTO(username).Result;
         return authorlist;
     }
     public async Task<List<CheepDTO>> GetCheepsFromFollowing(int pageNumber, string username)
     {
-        var cheeps = await _cheepRepository.GetCheepsFromFollowing(pageNumber, username);
+        var cheeps = await _followRepository.GetCheepsFromFollowing(pageNumber, username);
         return cheeps;
     }
 
     public async Task<List<CheepDTO>> GetCheepsFromAuthor(string author, int pageNumber)
     {
-        var result = await _cheepRepository.GetCheepsFromAuthor(pageNumber, author);
+        var result = await _followRepository.GetCheepsFromAuthor(pageNumber, author);
 
         // Debugging linje til logning
         Console.WriteLine($"Fetched {result.Count} cheeps for {author}");
@@ -67,10 +67,7 @@ public class FollowService : IFollowService
 
     public async Task<AuthorDTO> GetAuthorByName(string name)
     {
-        return await _cheepRepository.GetAuthorByName(name);
+        return await _followRepository.GetAuthorByName(name);
     }
-
-
-
 
 }
